Made some adjustments to enemy script
So I adjusted the enemy script a bit by adding in a raycast so we can make sure the character isn't hiding behind a rock and the enmy randomly sees them. Also put in a giant character but it deserves its own script. The current enemy script in there is closest to a giant enemy. Only able to see. next is the snake. Only able to sense movement. Most likely using a collider. lastly is human with both sensing movement and seeing!