Added touching finding AI for snake and a demo map with enemies.
If you are within a small range of the snakes (basically touching) you will get caught. Also added enemies throughout the map to playtest. Try to get the the sun where you can just run off the map! Still need to implement snakes communicating when you are found. Also need to light the mirror a little more because it is always dark. Maybe use a point light on character? Something to play with.