Fixed the snakespawning code a bit.
There was a bug so that if a instantiated snake gets too far away it will be destroyed and the count of snakes following you would decrement by 1 as it was suppose to. However, when the snake counter loops around to delete the first instantiated snake to make room for the new one it would decrement the numSnakesChasing you. However, if a snake got too far away then destroyed and decremented then when it looped back around in the array it would decrement again. Obviously not what we want. So now the snakeController script keeps track of it's index. When deleting a snake from being too far away, it will set its index in the array of snakes to null. Then when looping on the array to make room for the new ones it checks to see if the snake has already been deleted by checking to see if it is null first. If it is null then there is no need to decrement the amount of snakes chasing you by 1.