Found new tree to work with
Geez another late night... Started working a ton with the trees to raise the framerate up. Messed with a ton of graphics settings and made it so trees would dissapear if a little ways away from you. Did a ton of work but still ran into performance issues in certain sections. After much searching I found a sustainable tree that won't severly drop the frame rate. It comes at the cost of an uglier tree but I don't really think people will pay attention to it. As of now there is another demo scene in the scenes folder that I was testing on. i will take care of it all in the morning.