Snakes having changing color lights
This is to give users some conveyance on how the snake works since we can have any visual factors on the snake model. Make it a little easier but more fun (according to user testing). There is a bug that if the character moves really slowly in snake sphere they don't get noticed. I think I can fix it but I need a nap first