Found Giant's lag problem haven't fixed.
The problem is the radius of the character controller. No idea why increasing it takes such a toll on the CPU but it does. Anyways unfortunately I think we are going to have to shift it to a rigidbody and collider. It would probably be much more effecient because their character controller just blows! Obviously it's more work though so... But it has to be done.