Changed Human Animation
Altered the human animation so he has to take a step back before jumping over the lantern. Felt too cheap to not give any sort of break for using a lantern. Few steps back is minimal but it is something. It is kinda buggy though if you place multiple lanterns in a row