Bug fixes on AI
Fixed the length some AI chases you to make it shorter. Also fixed humanoid AI to make sure it only finds you at a specific angle and below and distance. There is still a bug where if there is a snake in your way it won't find you beause the linecast hits the snake.